Use setup tools and don't use relative paths (#19)
* Support loading views from /usr/share/steam-buddy

* Load images based on RESOURCE_DIR
This also fixes a bug. It was looking at ~/.local/share/banners/flathub
for images which are whitelisted.

* Move flathub to steam_buddy directory

* Added initial version setup.py file

* Fixed typo in setup.py

* Fixed adding views and images

* Fixed views image images not being added

* Added global path flathub whitelist check

* Added config and launcher files to setup.py

* Added LICENSE.md to package

* Fixed license file name in setup.py

* Fixed whitelist global image location

* Removed Makefile
It no longer worked and setuptools are now used

* Fixed steam-buddy location in systemd service

* Made views check more specific
I was checking if a directory named views was found in the current
directory and changing the working directory is not. Now it checks for a
specific view file instead.