Generate banners for new shortcuts if none is set (#102)
This uses the Pillow library to generate a simple banner with the title
of the game on it.
It currently uses a black background with a white font. It will
shorten titles if they don't fit on the banner.

This requires 2 dependencies:
- ttf-dejavu
- python-pillow

The package ttf-dejavu should already be a dependency for steam-buddy
(not sure if it is, though), since the authenticator uses it too.