Make flatpak games start reliably. Make command the complete path.
Some flatpaks need a StartDir set to start reliably. Currently if `dir` is not set, it's set to empty string. SuperTux fails to start until "StartDir" is set (didn't test any other).
Also set the command `cmd` to the full command to run the program. That way the `params` member can be used to set runtime parameters by the user.