Add FFmpeg default configuration.
Make it simpler for the user to add inputs as necessary.
By default use `x11grab` on input `:0` and `alsa` capture with `pulse`
input. This may not be the best solution but we want to make it
configurable in a way that the user can add as many inputs as he/she
pleases.

Audio and video codecs are not used by default but we add them as
configurables. They hold the entire line of configuration, meaning the
options for each one are set there as well.

For network streaming we hardcode `mpegts` as format (the only one
supported by SLS) with the option `flush_packets` set to `0` and for
local (recording) we hardcode it as `matroska` format and save the
`.mkv` file on users's home directory as `screen_yyyymmdd_hhmmss.mkv`.