Add Streaming Control and Configuration page (#152)
* Change SLS config to RESOURCE_DIR for now.

* Add FFmpeg default configuration.

Make it simpler for the user to add inputs as necessary.
By default use `x11grab` on input `:0` and `alsa` capture with `pulse`
input. This may not be the best solution but we want to make it
configurable in a way that the user can add as many inputs as he/she
pleases.

Audio and video codecs are not used by default but we add them as
configurables. They hold the entire line of configuration, meaning the
options for each one are set there as well.

For network streaming we hardcode `mpegts` as format (the only one
supported by SLS) with the option `flush_packets` set to `0` and for
local (recording) we hardcode it as `matroska` format and save the
`.mkv` file on users's home directory as `screen_yyyymmdd_hhmmss.mkv`.

* Add simple configuration for inputs.

It lists current inputs and allows to delete existing ones and add new
ones by adding the corresponding FFmpeg arguments

* Add configuration for Video and Audio codecs

Also add logic to handle strings and list for inputs, video and audio
codecs.

* Send recordings to dedicated directory

* Add Streaming Control page

* Fix stream.png borders