Install official proton if platform requires it
This will try to signal a running steam client if one of the installed
games from a platform needs a specific proton version. If the proton
version is an official one, then it will try to get steam client to
install it. Unofficial proton versions are managed by compat_tools
module.