Tweaks refactor to install official Proton versions (#193)
* test_scripts: Import modules instead of functions

* Refactor compat_tools for better integration

Preparation for better integration to config and shortcuts
Add mechanisms to install official compatibility tools

* Refactor steam_config for better integration

* Refactor shortcuts for better integration

* Improve linting on platform handlers

* Remove load_shortcuts in favor of classes

* Install official proton if platform requires it

This will try to signal a running steam client if one of the installed
games from a platform needs a specific proton version. If the proton
version is an official one, then it will try to get steam client to
install it. Unofficial proton versions are managed by compat_tools
module.