Complete system-info compatibility with Steam System Information and ProtonDB
Extends system-info so it produces output that is 100% compatible with what Steam System Information produces, including the inconsistent double and single character spacing. The resulting output is also compatible with ProtonDB. Best efforts are made to automatically hunt for an accessible DISPLAY. 

Unlike Steam System Information on Linux, this system-info will correctly identify touch devices and correctly count HDDs and SSDs.
  - https://github.com/ValveSoftware/steam-for-linux/issues/5881
  - https://github.com/ValveSoftware/steam-for-linux/issues/8412

Motivation for creating patch are that system-info might be more useful a general information gathering tool for users reporting issues and that system-info might now be suitable for integrating in the Chimera webapp, allowing ChimeraOS users an easy way to add their consoles to their ProtonDB profile.
  - https://github.com/ChimeraOS/chimera/issues/231

Make the DISPLAY accessibility checks work on Wayland/Xwayland


If form factor is empty, report it as unknown.


Make the DISPLAY accessibility tests a noop


Correct typo